Fix ws.get_extra_info('peername') calls for IPV6
If the websocket is connected over IPV6, ws.get_extra_info('peername') will return a 4-tuple instead of a 2-tuple. This was causing lines where a 2-tuple was always expected to break. I've added a function which does the ws.get_extra_info('peername') call and always returns a 2-tuple, which can be used in-place of the old ws.get_extra_info('peername') calls.